fix integer overflow which triggers an fpe in sprintf_number

+++ src/nem/format/readers.h
@@ -23,7 +23,7 @@
 
 uint16_t sprintf_hex(char *dst, uint16_t maxLen, const uint8_t *src, uint16_t dataLength, uint8_t reverse);
 uint16_t sprintf_ascii(char *dst, uint16_t maxLen, const uint8_t *src, uint16_t dataLength);
-uint16_t sprintf_number(char *dst, uint16_t maxLen, uint64_t value);
+int sprintf_number(char *dst, uint16_t maxLen, uint64_t value);
 uint16_t sprintf_mosaic(char *dst, uint16_t maxLen, const uint8_t *mosaic, uint16_t dataLength);
 uint16_t sprintf_token(char* dst, uint16_t len, uint64_t amount, uint8_t divisibility, char* token);
 uint16_t sprintf_hex2ascii(char *dst, uint16_t maxLen, const uint8_t *src, uint16_t dataLength);
